feat: move wallet and token icons to its own folder
