+++ src/components/Icons/index.ts
@@ -3,20 +3,19 @@ export * from './ChevronLeftIcon';
 export * from './CopyIcon';
 export * from './DeleteIcon';
 export * from './DropIcon';
-export * from './HereLogoIcon';
-export * from './HereWalletIcon';
 export * from './ImageIcon';
 export * from './KeyIcon';
 export * from './LinkIcon';
 export * from './LogoNear';
 export * from './MenuIcon';
-export * from './MyNearWalletIcon';
 export * from './NearIcon';
 export * from './NFTIcon';
 export * from './POAPNftIcon';
 export * from './SignOutIcon';
 export * from './StarIcon';
 export * from './TicketIcon';
-export * from './TokenEthIcon';
+export * from './tokens/TokenEthIcon';
 export * from './TwitterLogoIcon';
 export * from './UncheckedIcon';
+export * from './wallets/HereWalletIcon';
+export * from './wallets/MyNearWalletIcon';
