+++ src/components/Icons/tokens/TokenEthIcon.tsx
