+++ src/components/Icons/wallets/MyNearWalletIcon.tsx
